refactor: move client builder to IQOptionBuilder
