+++ src/lib.rs
@@ -1,7 +1,11 @@
+mod builder;
 mod client;
 mod credentials;
 mod enviroment;
+mod session;
 
+pub use builder::IQOptionBuilder;
 pub use client::IQOptionClient;
 pub use credentials::Credentials;
 pub use enviroment::Enviroment;
+pub use session::Session;
