+++ src/session.rs
@@ -1,0 +1,5 @@
+/// Data for a session of the IQOption API.
+pub struct Session {
+    /// Session ID.
+    pub ssid: String,
+}
